Noticket: fix PMD warnings
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@14149 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/ejbca-entity/src/org/cesecore/dbprotection/ProtectedDataConfiguration.java
@@ -214,7 +214,7 @@ public final class ProtectedDataConfiguration {
 		        					// Data is base64 encoded byte[] so decode it
 		        					keydata = Base64.decode(data.getBytes());	        						
 	        					}
-	        					final CryptoToken token = CryptoTokenFactory.createCryptoToken(classname, properties, keydata, Integer.valueOf(keyid).intValue());
+	        					final CryptoToken token = CryptoTokenFactory.createCryptoToken(classname, properties, keydata, Integer.valueOf(keyid));
 	        					// We must activate the token as well (if not using a default pwd of course, in which case we assume the tokenpin property is not set)
 	        					final String pin = conf.getString(pinstr+i);
 	        					try {
